Rename file and require() -> import
